pyproject.toml
[tool.poetry]
name = "dls-emailer"
version = "0.1.0"
description = ""
authors = ["Morten Bendeke <mbendeke@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
